ProyectoFruta/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
opencv-python==4.12.0.88
onnxruntime==1.22.1
numpy==2.2.6
Pillow==11.3.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic==2.5.0
aiofiles==23.2.1
